Create 15_improving_visa_process.md
Getting Visa Approval (Seamless visa process)  is a major issue highlighted during the 2025 DISC unconference as a barrier that is contributing to the lack of inclusiveness and accessibility in Scientific events. In a bid to mitigate this harm, we wrote a chapter that is meant to serve as a guide in addressing this problem.